+++ src/test/java/com/synopsys/integration/detect/workflow/blackduck/developer/RapidModeGenerateJsonOperationTest.java
@@ -29,7 +29,7 @@ public class RapidModeGenerateJsonOperationTest {
 
         File tempDir = tempPath.toFile();
         File scanDir = new File(tempDir, "scan");
-        DirectoryOptions directoryOptions = new DirectoryOptions(null, null, null, scanDir.toPath(), null, null);
+        DirectoryOptions directoryOptions = new DirectoryOptions(null, null, null, scanDir.toPath(), null, null, null);
         DetectRunId detectRunId = new DetectRunId("testId");
         DirectoryManager directoryManager = new DirectoryManager(directoryOptions, detectRunId);
         RapidModeGenerateJsonOperation op = new RapidModeGenerateJsonOperation(gson, directoryManager);